pyproject.toml
[project]
name = "pkdb-codex"
version = "0.1.0"
description = "Pharmacological data management platform (first draft)"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "python-jose>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
mongo = ["pymongo>=4.7.0"]
test = ["pytest>=8.2.0", "httpx>=0.27.0"]

[tool.pytest.ini_options]
addopts = "-q"
